requirements.txt
torch>=1.12.0
torchvision>=0.13.0
lpips>=0.1.4
numpy>=1.21.6
pillow>=8.0.0
onnx>=1.15.0
onnxruntime>=1.15.0